requirements.txt
pytest==8.0.0
pytest-asyncio==0.23.5
sqlalchemy==2.0.25
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
bcrypt==4.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
aiosqlite==0.21.0
starlette==0.36.3
celery>=5.3.0
redis>=4.5.0 